Refactor(files): Reorganize domain DTOs and enums
