+++ libs/common/src/lib/files/domain/dtos/file.dto.ts
@@ -1,4 +1,4 @@
-import { FileStatus } from './file-status.enum';
+import { FileStatus } from '../enums/file-status.enum';
 
 export class FileDto {
   id!: string;
