+++ libs/common/src/lib/files/domain/dtos/folder.dto.ts
