+++ libs/common/src/lib/files/domain/dtos/index.ts
@@ -1,0 +1,3 @@
+export * from './file.dto';
+export * from './folder.dto';
+export * from './permission.dto';
