+++ libs/common/src/lib/files/domain/dtos/permission.dto.ts
@@ -1,5 +1,4 @@
-import { PermissionRole } from './permission-role.enum';
-import { PermissionType } from './permission-type.enum';
+import { PermissionRole, PermissionType } from '../enums';
 
 export class PermissionDto {
   subjectId!: string; // User ID or Group ID
