+++ libs/common/src/lib/files/domain/enums/index.ts
@@ -1,0 +1,3 @@
+export * from './file-status.enum';
+export * from './permission-role.enum';
+export * from './permission-type.enum';
