+++ libs/common/src/lib/files/index.ts
@@ -1,6 +1,2 @@
-export * from './file-status.enum';
-export * from './file.dto';
-export * from './folder.dto';
-export * from './permission-role.enum';
-export * from './permission-type.enum';
-export * from './permission.dto';
+export * from './domain/dtos';
+export * from './domain/enums';
